Proxy hotfix
